LayoutTests/inspector/console/console-log-syntax-error-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token ')'
Tests that syntax errors are logged into console and doesn't cause browser crash.

syntax-error.js:1SyntaxError: Unexpected token ')'
